commons fields in abstract connector
