+++ src/main/java/fr/edf/jenkins/plugins/windows/connector/MicroServiceJNLPConnector.groovy
@@ -2,12 +2,20 @@ package fr.edf.jenkins.plugins.windows.connector
 
 import java.time.Instant
 
+import javax.net.ssl.SSLContext
+import javax.net.ssl.X509TrustManager
+
 import org.apache.commons.lang.exception.ExceptionUtils
+import org.apache.http.client.HttpClient
+import org.apache.http.impl.client.CloseableHttpClient
+import org.apache.http.impl.client.HttpClientBuilder
 import org.jenkinsci.Symbol
+import org.kohsuke.stapler.DataBoundConstructor
 
 import fr.edf.jenkins.plugins.windows.WindowsHost
 import fr.edf.jenkins.plugins.windows.WindowsUser
 import fr.edf.jenkins.plugins.windows.agent.WindowsComputer
+import fr.edf.jenkins.plugins.windows.http.MicroserviceHttpClient
 import fr.edf.jenkins.plugins.windows.winrm.WinRMCommandException
 import hudson.Extension
 import hudson.model.Descriptor
@@ -18,13 +26,14 @@ import hudson.slaves.SlaveComputer
 
 class MicroServiceJNLPConnector extends WindowsComputerConnector {
 
-    private String jenkinsUrl;
-    private Integer port;
-    private boolean useHttps;
-    private boolean ignoreCertificate;
-    private String credentialsId;
-    private Integer connectionTimeout;
-    private Integer readTimeout;
+    private MicroserviceHttpClient client
+
+    @DataBoundConstructor
+    MicroServiceJNLPConnector(Boolean useHttps, Boolean disableCertificateCheck, Integer port,
+    String credentialsId, Integer connectionTimeout, Integer readTimeout,
+    Integer agentConnectionTimeout, String jenkinsUrl) {
+        super(jenkinsUrl, port, useHttps, disableCertificateCheck, credentialsId, connectionTimeout, readTimeout, agentConnectionTimeout)
+    }
 
     @Override
     protected ComputerLauncher createLauncher(WindowsHost host, WindowsUser user) {
